part2_crewai/requirements.txt
# Part 2 — CrewAI edition
# No LangChain agent layer — pure Python orchestration
# Windows + Python 3.13 compatible

python-dotenv
pytest
pytest-asyncio
